AD-HOC feat (Ansible): Improve documented CLI arguments
Currently the CLI arguments appear to be expressed in a match and
replace form. However, the template in which these CLI flags are
rendered uses the following syntax:

  {% for flag in osquery_flags %}
  {{ flag }}
  {% endfor %}

This means that the flags, as documented in the file, are expressed
improperly.

This commit modifies the documentation to note how the flags should be
expressed in the file.